+++ protocol/__init__.py
@@ -1,0 +1,8 @@
+"""
+The `protocol` module provides data model classes related to the CAPTCHA API.
+
+These classes are designed for use by both frontend and backend applications.
+"""
+from .tiles import TilesResponse
+
+__all__ = ["TilesResponse"]
